src/requirements.txt
altgraph==0.17.2
beautifulsoup4==4.10.0
future==0.18.2
greenlet==1.1.2
pefile==2021.9.3
Pillow==8.4.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pywin32-ctypes==0.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.28
tkinterdnd2==0.3.0
